Add scaled delay.
This will help to keep the animation within a reasonable amount of time
no matter how big the dataset grows too.

We 'normalize' the value by dividing by the length of the dataset and
limit the delay to be 1000 or less ms. Therefore, no matter the size of
the dataset, the total transition animation time will always be 1.5s or
less.