Implement key function.
Create variable 'key' to reference the 'key' property of each data
point.

Situation: When we remove a value from the beginning of the dataset
array this causes _all_ values to be re-binded to a new DOM element
since the index position of each element is changed. This causes the
bars to shift to the right.[Needs more explanation...]

Solution: Instead of binding a data point to a DOM element taking into
account the index position of that data point in dataset, we force D3
to use the 'key' property of a data point instead.

--> As long as all the keys are unique, then this should work.